+++ CHANGES.md
@@ -2,8 +2,34 @@ Change History
 ==============
 
 
-1.1.0 for MC 1.20.2 (2024-04-13)
---------------------------------
+1.2.1 for Minecraft 1.20.2 with Forge 48.1.0 (2024-08-18)
+---------------------------------------------------------
+
+Bug fixes:
+
+- Prevent possible crash on bad start-up.
+
+
+1.2.0 for Minecraft 1.20.2 with Forge 48.1.0 (Unreleased)
+---------------------------------------------------------
+
+New features:
+
+- Added the "collector.mc_dimension_tick_errors" setting to control how to handle inconsistent dimension ticks from misbehaved mods. The new default behavior is to log a debug message rather than crash.
+
+Bug fixes:
+
+- Do not warn when the mod is not installed on the client.
+- Support inconsistent dimension ticks from misbehaved mods.
+- Support multithreaded dimension ticks.
+
+Miscellaneous:
+
+- Minor documentation.
+
+
+1.1.0 for Minecraft 1.20.2 with Forge 48.1.0 (2024-04-13)
+---------------------------------------------------------
 
 New features:
 
@@ -19,9 +45,3 @@ Miscellaneous:
 - Added "HACKING.md".
 - Added "dashboards.md".
 - Added "metrics.md".
-
-
-1.0.0 for MC 1.20.1 (2023-08-10)
---------------------------------
-
-- [Issue #12](https://github.com/cpburnz/minecraft-prometheus-exporter/issues/12): Support MC 1.20.1.